requirements.txt
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
